Replace withOpenInvoice with OpenInvoiceContainer helper and add relevant tests.
